requirements.txt
pyserial-3.4
cx-Freeze-6.0
